requirements.txt
# Core dependencies
lancedb>=0.5.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
pymupdf>=1.23.0
unstructured>=0.10.0
pdfplumber>=0.9.0
camelot-py[cv]>=0.11.0

# LLM and AI
llama-cpp-python>=0.2.0
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Web framework (optional)
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
click>=8.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 